[ocf-servers] Add an environmental sensor OIC resource.
It uses noble to connect via BLE to the environmental sensor to get data.
There are temperature, humidity, air pressure and UV index sensor data that
the resource represents. To use, launch the nodejs application in the
homegateway with BLE capability. It has been tested on MinnowBoard Turbot &
Galileo with a CSR 4.0 BLE USB dongle. Edison is not yet supported.

Signed-off-by: Kien Dinh <kien.t.dinh@intel.com>